Water Bottle Index
One issue plaguing water use research is the inability to appropriately scale water use by individuals. For example, the dominant water use unit for the Southwest US is an acre-foot. While it may be helpful to know an acre-foot is the amount of water required to cover an acre of land with one foot of water, it is not intuitive nor easy for someone to conceptualize how much water that is. To create an easier methodological tool for individuals I create the water bottle index, which defines individual water consumption by the amount of water bottles.